requirements.txt
streamlit
llama_index
llama-index-core
matplotlib
python-dotenv
flask
docx2txt
python-docx

